add some more tests.

+++ Imaging/ColorHsv.cs
@@ -288,11 +288,11 @@ namespace Imaging
         /// <returns>A tuple containing the RGBA values.</returns>
         public void ColorToHsv(int color)
         {
-            // Extract ARGB values from the integer
-            A = (color >> 24) & 0xFF; // Alpha
-            R = (color >> 16) & 0xFF; // Red
-            G = (color >> 8) & 0xFF;  // Green
-            B = color & 0xFF;         // Blue
+            // Extract ARGB values from the unsigned integer
+            A = (byte)((color >> 24) & 0xFF); // Alpha
+            R = (byte)((color >> 16) & 0xFF); // Red
+            G = (byte)((color >> 8) & 0xFF);  // Green
+            B = (byte)(color & 0xFF);         // Blue
 
             // Calculate max and min values for HSV conversion
             var max = Math.Max(R, Math.Max(G, B));
@@ -306,6 +306,7 @@ namespace Imaging
             // Update the hex representation of the color
             GetHex();
         }
+
 
         /// <summary>
         ///     Hexadecimals to color.
